Migrate ButtonGroup to Styled Components (#342)
